+++ implementation/src/main/java/com/avioconsulting/mule/testing/muleinterfaces/wrappers/EventWrapper.groovy
@@ -10,20 +10,28 @@ interface EventWrapper {
     EventWrapper withVariable(String variableName,
                               Object value)
 
+    EventWrapper withVariable(String variableName,
+                              Object value,
+                              String mediaType)
+
     EventWrapper withNewAttributes(attributes)
 
     EventWrapper withNewPayload(Object payload,
+                                ConnectorInfo connectorInfo,
                                 String mediaType)
 
     EventWrapper withSoapPayload(String xmlPayload,
+                                 ConnectorInfo connectorInfo,
                                  Map attributes)
 
     EventWrapper withNewPayload(Object payload,
                                 String mediaType,
+                                ConnectorInfo connectorInfo,
                                 Map attributes)
 
     EventWrapper withNewStreamingPayload(String payload,
                                          String mediaType,
                                          Map attributes,
+                                         ConnectorInfo connectorInfo,
                                          boolean useRepeatableStream)
 }