+++ implementation/src/main/java/com/avioconsulting/mule/testing/muleinterfaces/wrappers/InterceptEventWrapperImpl.groovy
@@ -2,18 +2,23 @@ package com.avioconsulting.mule.testing.muleinterfaces.wrappers
 
 class InterceptEventWrapperImpl extends
         EventWrapperImpl {
+    private final Map<String, Object> variableOverrides
+
     InterceptEventWrapperImpl(Object nativeEvent,
                               Object runtimeBridgeMuleSide) {
         super(nativeEvent,
               runtimeBridgeMuleSide)
+        variableOverrides = [:]
     }
 
     private InterceptEventWrapperImpl(Object existingNativeEvent,
                                       MessageWrapper newMessageWrapper,
-                                      Object runtimeBridgeMuleSide) {
+                                      Object runtimeBridgeMuleSide,
+                                      Map<String, Object> variableOverrides) {
         super(existingNativeEvent,
               newMessageWrapper,
               runtimeBridgeMuleSide)
+        this.variableOverrides = variableOverrides
     }
 
     @Override
@@ -22,16 +27,48 @@ class InterceptEventWrapperImpl extends
         // mocks can't return new events, they have to mutate, so we'll mutate the message
         // inside the underlying event
         this.nativeEvent.message(muleMsg)
-        // you can't read .message of the mock native event and see the new message (it will be the original one)
+        // you can't read .message of the mock native event (see DefaultInterceptionEvent, in vs. out)
+        // and see the new message (it will be the original one)
         // so we have to pass it along ourselves to the private constructor above
         return new InterceptEventWrapperImpl(this.nativeEvent,
                                              newMessage,
-                                             runtimeBridgeMuleSide)
+                                             runtimeBridgeMuleSide,
+                                             variableOverrides)
+    }
+
+    @Override
+    Object getVariable(String variableName) {
+        // just as described in withNewMessage above, you can't read new flowVars, so if we change, we track it
+        variableOverrides.get(variableName) ?: super.getVariable(variableName)
     }
 
     @Override
     EventWrapper withVariable(String variableName,
                               Object value) {
-        throw new Exception('Have not implemented this yet!')
+        this.nativeMuleEvent.addVariable(variableName,
+                                         value)
+        // just as described in withNewMessage above, you can't read new flowVars, so if we change, we track it
+        def newOverrides = this.variableOverrides + [(variableName): value]
+        return new InterceptEventWrapperImpl(this.nativeEvent,
+                                             this.message,
+                                             runtimeBridgeMuleSide,
+                                             newOverrides)
+    }
+
+    @Override
+    EventWrapper withVariable(String variableName,
+                              Object value,
+                              String mediaType) {
+        def resolvedMediaType = runtimeBridgeMuleSide.getDataType(value,
+                                                                  mediaType)
+        this.nativeMuleEvent.addVariable(variableName,
+                                         value,
+                                         resolvedMediaType)
+        // just as described in withNewMessage above, you can't read new flowVars, so if we change, we track it
+        def newOverrides = this.variableOverrides + [(variableName): value]
+        return new InterceptEventWrapperImpl(this.nativeEvent,
+                                             this.message,
+                                             runtimeBridgeMuleSide,
+                                             newOverrides)
     }
 }