+++ implementation/src/main/java/com/avioconsulting/mule/testing/transformers/ClosureCurrier.groovy
@@ -3,10 +3,22 @@ package com.avioconsulting.mule.testing.transformers
 import com.avioconsulting.mule.testing.muleinterfaces.wrappers.ConnectorInfo
 import com.avioconsulting.mule.testing.muleinterfaces.wrappers.EventWrapper
 
-interface ClosureCurrier<T extends ConnectorInfo> {
-    boolean isOnlyArgumentToBeCurried(Closure closure)
-
+class ClosureCurrier<T extends ConnectorInfo> {
     Closure curryClosure(Closure closure,
                          EventWrapper muleEvent,
-                         T messageProcessor)
+                         T messageProcessor) {
+        closure = doCurry(closure,
+                          muleEvent)
+        doCurry(closure,
+                messageProcessor)
+    }
+
+    private static Closure doCurry(Closure closure,
+                                   Object object) {
+        def parameterTypes = closure.parameterTypes
+        if (parameterTypes.any() && parameterTypes.last().isAssignableFrom(object.getClass())) {
+            closure = closure.rcurry(object)
+        }
+        closure
+    }
 }