+++ implementation/src/main/java/com/avioconsulting/mule/testing/transformers/OutputTransformer.groovy
@@ -1,8 +1,10 @@
 package com.avioconsulting.mule.testing.transformers
 
+import com.avioconsulting.mule.testing.muleinterfaces.wrappers.ConnectorInfo
 import com.avioconsulting.mule.testing.muleinterfaces.wrappers.EventWrapper
 
 interface OutputTransformer {
     EventWrapper transformOutput(input,
-                                 EventWrapper originalMuleEvent)
+                                 EventWrapper originalMuleEvent,
+                                 ConnectorInfo connectorInfo)
 }