+++ implementation/src/main/java/com/avioconsulting/mule/testing/transformers/json/output/JacksonOutputTransformer.groovy
@@ -1,5 +1,6 @@
 package com.avioconsulting.mule.testing.transformers.json.output
 
+import com.avioconsulting.mule.testing.muleinterfaces.wrappers.ConnectorInfo
 import com.avioconsulting.mule.testing.muleinterfaces.wrappers.EventWrapper
 import com.avioconsulting.mule.testing.transformers.OutputTransformer
 import com.fasterxml.jackson.databind.ObjectMapper
@@ -14,7 +15,8 @@ class JacksonOutputTransformer implements
     }
 
     EventWrapper transformOutput(Object input,
-                                 EventWrapper originalMuleEvent) {
+                                 EventWrapper originalMuleEvent,
+                                 ConnectorInfo connectorInfo) {
         def jsonString = getJsonOutput(input)
         def messageProps = [
                 'http.status': '200'
@@ -22,6 +24,7 @@ class JacksonOutputTransformer implements
         originalMuleEvent.withNewStreamingPayload(jsonString,
                                                   'application/json',
                                                   messageProps,
+                                                  connectorInfo,
                                                   this.useRepeatableStream)
     }